transfer and restructure Euclidean function.
